upgrade scripts

+++ packages/devtools/scripts/static-build.js
@@ -6,7 +6,7 @@ process.env.NODE_ENV = 'production';
 // Makes the script crash on unhandled rejections instead of silently
 // ignoring them. In the future, promise rejections that are not handled will
 // terminate the Node.js process with a non-zero exit code.
-process.on('unhandledRejection', err => {
+process.on('unhandledRejection', (err) => {
   throw err;
 });
 
@@ -15,7 +15,7 @@ const path = require('path');
 // Ensure environment variables are read.
 require('../config/env');
 
-const requireRelative = relativePath => require(path.join(__dirname, relativePath));
+const requireRelative = (relativePath) => require(path.join(__dirname, relativePath));
 
 const webpack = require('webpack');
 const fs = require('fs-extra');
@@ -28,12 +28,11 @@ const formatWebpackMessages = require('react-dev-utils/formatWebpackMessages');
 const measureFileSizesBeforeBuild = FileSizeReporter.measureFileSizesBeforeBuild;
 const printFileSizesAfterBuild = FileSizeReporter.printFileSizesAfterBuild;
 const configureWebpackClient = requireRelative('../webpack/client').configure;
-const merge = require('lodash').merge;
 
 // First, read the current file sizes in build directory.
 // This lets us display how much they changed later.
 measureFileSizesBeforeBuild(paths.appBuild)
-  .then(previousFileSizes => {
+  .then((previousFileSizes) => {
     // Remove all content but keep the directory so that
     // if you're in it, you don't end up in Trash
     fs.emptyDirSync(paths.appBuild);
@@ -60,7 +59,7 @@ measureFileSizesBeforeBuild(paths.appBuild)
       printFileSizesAfterBuild(stats, previousFileSizes, paths.appBuild);
       console.log();
     },
-    err => {
+    (err) => {
       console.log(chalk.red('Failed to compile.\n'));
       console.log((err.message || err) + '\n');
       process.exit(1);
@@ -72,7 +71,7 @@ function build(previousFileSizes) {
 
   const localBuildConfig = fs.existsSync(paths.localBuildConfig) ? require(paths.localBuildConfig) : {};
 
-  const buildConfig = merge(globalBuildConfig, localBuildConfig);
+  const buildConfig = { ...globalBuildConfig, ...localBuildConfig };
 
   let clientConfig = !!buildConfig.client && configureWebpackClient(buildConfig.client);
 
@@ -113,9 +112,13 @@ function build(previousFileSizes) {
 }
 
 function copyPublicFolder() {
+  if (!fs.existsSync(paths.appPublic)) {
+    return;
+  }
+
   fs.copySync(paths.appPublic, paths.appBuildPublic, {
     dereference: true,
-    filter: file => file !== paths.appHtml
+    filter: (file) => file !== paths.appHtml
   });
 }
